requirements.txt
torch
torchvision
gradio
pillow
huggingface-hub